backend/src/requirements.txt
Flask-PyMongo==2.3.0
pymongo==4.3.3